Finish util lab.
In this lab, the most work is dirty work. The most interesting part is
to use pipe as a pipeline to achieve concurrency. The idea is simple.
Just use `read` to block perfectly. This is the funny thing. However,
`find` and `xargs` is a dirty work, the idea is super simply. But it is
tedious for debugging.